建议类 clean code 修改
# message auto-generated for no-merge-commit merge:
merge main into master

建议类 clean code 修改

Created-by: msprofDeveloper

Match-id-2f31c94877a479640e3f490d3444af660521ce62

+++ analysis/csrc/viewer/database/finals/npu_module_mem_processor.cpp
@@ -62,7 +62,7 @@ NpuModuleMemProcessor::ProcessedDataFormat NpuModuleMemProcessor::FormatData(con
     }
     for (auto &row: oriData) {
         std::tie(data.moduleId, data.syscnt, data.totalSize, data.deviceType) = row;
-        HPFloat timestamp{GetTimeFromSyscnt(data.syscnt, params)};
+        HPFloat timestamp{GetTimeFromSyscnt(static_cast<uint64_t>(data.syscnt), params)};
         processedData.emplace_back(
             data.moduleId, GetLocalTime(timestamp, timeRecord).Uint64(), data.totalSize, deviceId);
     }
